feat: refactor chat page route
